requirements.txt
Flask
cloudant
gunicorn
python-dotenv
